+++ src/features/modals/client-modal/client-edit-modal.tsx
@@ -6,16 +6,16 @@ import { useEffect } from 'react';
 
 import FormInput from '@/components/form/form-input';
 import FormTextarea from '@/components/form/form-textarea';
+import FormSwitch from '@/components/form/form-switch';
 import { Button } from '@/components/ui/button';
 import { Form } from '@/components/ui/form';
+import { FormLabel } from '@/components/ui/form';
 import MaterialSymbolsContentCopy from '~icons/material-symbols/content-copy';
 
 import useClientInfo from '@/services/hooks/client/use-client-info';
 import useUpdateClient from '@/services/hooks/client/use-update-client';
 import useDeleteClient from '@/services/hooks/client/use-delete-client';
-import FormSwitch from '@/components/form/form-switch';
 import { useModalContext } from '@/services/providers/modal-provider';
-import { FormLabel } from '@/components/ui/form';
 
 import { z } from '@/utils/zod';
 
@@ -77,69 +77,63 @@ const Component = ({ client_reference }: Props) => {
     };
 
     const onCopy = async (formData: z.infer<typeof formSchema>) => {
-        navigator.clipboard.writeText(formData.secret)
-    }
+        navigator.clipboard.writeText(formData.secret);
+    };
 
-    if (!data) return null
+    if (!data) return null;
 
     return (
         <Form {...form}>
-            <form
-                onSubmit={(e) => e.preventDefault()}
-                className="flex flex-col gap-6"
-            >
-                <div className="flex w-full flex-col gap-6">
-                    <div className="flex w-full flex-col gap-6">
-                        <FormInput
-                            name="name"
-                            label="Назва додатка"
-                            placeholder="Введіть назву застосунка"
-                            type="string"
-                        />
-                        <FormTextarea
-                            name="description"
-                            label="Опис"
-                            placeholder="Залиште опис до застосунку"
-                        />
-                        <FormInput
-                            name="endpoint"
-                            label="Посилання переспрямування"
-                            placeholder="https://example.com/"
-                            type="string"
-                        />
-                        <div>
-                            <FormLabel>Секрет</FormLabel>
-                            <div className="flex items-end flex-row w-full gap-2">
-                                <FormInput
-                                    name="secret"
-                                    placeholder="h1Kk@--H3l1o1tsl0rgoN- ..."
-                                    disabled={true}
-                                    type="password"
-                                    className="w-full"
-                                />
-                                <Button 
-                                    variant="secondary"
-                                    onClick={form.handleSubmit(onCopy)}
-                                >
-                                    <MaterialSymbolsContentCopy />
-                                    Скопіювати
-                                </Button>
-                            </div>
+            <form onSubmit={(e) => e.preventDefault()} className="flex flex-col gap-6">
+                <div className="flex flex-col gap-6 w-full">
+                    <FormInput
+                        name="name"
+                        label="Назва додатка"
+                        placeholder="Введіть назву застосунка"
+                        type="string"
+                    />
+                    <FormTextarea
+                        name="description"
+                        label="Опис"
+                        placeholder="Залиште опис до застосунку"
+                    />
+                    <FormInput
+                        name="endpoint"
+                        label="Посилання переспрямування"
+                        placeholder="https://example.com/"
+                        type="string"
+                    />
+                    <div>
+                        <FormLabel>Секрет</FormLabel>
+                        <div className="flex items-end gap-2 w-full">
+                            <FormInput
+                                name="secret"
+                                placeholder="h1Kk@--H3l1o1tsl0rgoN- ..."
+                                disabled
+                                type="password"
+                                className="w-full"
+                            />
+                            <Button 
+                                variant="secondary"
+                                onClick={form.handleSubmit(onCopy)}
+                            >
+                                <MaterialSymbolsContentCopy />
+                                Скопіювати
+                            </Button>
                         </div>
-                        <FormSwitch
-                            name="revoke_secret"
-                            label="Перестворити секрет"
-                            className="w-full"
-                        />
                     </div>
-                    <div className="grid w-full grid-cols-2 gap-8">
+                    <FormSwitch
+                        name="revoke_secret"
+                        label="Перестворити секрет"
+                        className="w-full"
+                    />
+                    <div className="grid grid-cols-2 gap-8 w-full">
                         <Button
                             variant="accent"
                             onClick={form.handleSubmit(onUpdate)}
                             type="submit"
-                            disabled={false}
                         >
-                            {clientInfoLoading || updateClientLoading || deleteClientLoading && (
+                            {(updateClientLoading) && (
                                 <span className="loading loading-spinner"></span>
                             )}
                             Оновити
@@ -148,9 +142,8 @@ const Component = ({ client_reference }: Props) => {
                             type="button"
                             variant="destructive"
                             onClick={onDelete}
-                            disabled={false}
                         >
-                            {clientInfoLoading || updateClientLoading || deleteClientLoading && (
+                            {(deleteClientLoading) && (
                                 <span className="loading loading-spinner"></span>
                             )}
                             Видалити
